requirements.txt
astroid==2.1.0
autopep8==1.4.3
bcrypt==3.1.6
cffi==1.12.1
Click==7.0
colorama==0.4.1
Flask==1.0.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mongoengine==0.16.3
pycodestyle==2.5.0
pycparser==2.19
pylint==2.2.2
pymongo==3.7.2
python-dotenv==0.10.1
six==1.12.0
Werkzeug==0.14.1
wrapt==1.11.1
